refactor: move `IntOrFloat` to `shared` module
